Update the vsp implementation code
